CRUD for Story Entity.

+++ src/organizations/organizations.service.ts
@@ -4,6 +4,7 @@ import {
   NotFoundException,
 } from "@nestjs/common";
 
+import { SharedProjectsService } from "../shared-projects/shared-projects.service";
 import { CreateOrganizationDto } from "./dtos/create-organization.dto";
 import { UpdateOrganizationDto } from "./dtos/update-organization.dto";
 import { OrganizationsRepository } from "./organizations.repository";
@@ -11,7 +12,8 @@ import { OrganizationsRepository } from "./organizations.repository";
 @Injectable()
 export class OrganizationsService {
   constructor(
-    private readonly organizationsRepository: OrganizationsRepository
+    private readonly organizationsRepository: OrganizationsRepository,
+    private readonly sharedProjectsService: SharedProjectsService
   ) {}
 
   async createOrganization(data: CreateOrganizationDto, ownerId: string) {
@@ -35,12 +37,16 @@ export class OrganizationsService {
     return this.organizationsRepository.findOrgMember(userId, organizationId);
   }
 
+  findOrganizationById(organizationId: string) {
+    return this.organizationsRepository.findOrganizationById(organizationId);
+  }
+
   async updateOrganization(
     organizationId: string,
     data: UpdateOrganizationDto
   ) {
     const organizationToUpdate =
-      await this.organizationsRepository.findOrganizationById(organizationId);
+      await this.findOrganizationById(organizationId);
     if (!organizationToUpdate)
       throw new NotFoundException("Organization does not exists.");
 
@@ -60,7 +66,21 @@ export class OrganizationsService {
     );
   }
 
-  deleteOrganization(organizationId: string) {
+  async deleteOrganization(userId: string, organizationId: string) {
+    const organizationToUpdate =
+      await this.findOrganizationById(organizationId);
+    if (!organizationToUpdate)
+      throw new NotFoundException("Organization does not exists.");
+
+    const projects = await this.sharedProjectsService.findProjectsByUserAndOrg(
+      userId,
+      organizationId
+    );
+    if (projects?.length > 0)
+      throw new BadRequestException(
+        "Unable to delete organization, there are projects found in this organization."
+      );
+
     return this.organizationsRepository.deleteOrganization(organizationId);
   }
 }
